55: Add performance tests http://code.google.com/p/sqljet/issues/detail?id=55
deleteAll

git-svn-id: http://svn.svnkit.com/repos/sqljet@539 06c6b4df-3a75-4b2a-9d74-3e57e8079485

+++ trunk/sqljet-test/src/org/tmatesoft/sqljet/benchmarks/SqlJetBenchmark.java
@@ -107,7 +107,11 @@ public class SqlJetBenchmark extends AbstractBenchmark {
                 final ISqlJetCursor c = table.open();
                 try {
                     do {
-                        c.delete();
+                        if (!c.isNull(0)) {
+                            c.delete();
+                        } else {
+                            c.next();
+                        }
                     } while (!c.eof());
                 } finally {
                     c.close();
